+++ platformio/__init__.py
@@ -1,17 +1,18 @@
 # Copyright (C) Ivan Kravets <me@ikravets.com>
 # See LICENSE for details.
 
-VERSION = (0, 6, 0)
+VERSION = (0, 7, 0)
 __version__ = ".".join([str(s) for s in VERSION])
 
 __title__ = "platformio"
 __description__ = ("A cross-platform code builder and library manager")
-__url__ = "https://github.com/ivankravets/platformio"
+__url__ = "http://platformio.ikravets.com"
 
 __author__ = "Ivan Kravets"
 __email__ = "me@ikravets.com"
 
-__license__ = "MIT Licence"
+__license__ = "MIT License"
 __copyright__ = "Copyright (C) 2014 Ivan Kravets"
 
-__pkgmanifesturl__ = "http://platformio.ikravets.com/packages/manifest.json"
+__apiurl__ = "http://api.platformio.ikravets.com"
+__pkgmanifesturl__ = "http://dl.platformio.ikravets.com/packages/manifest.json"