+++ platformio/exception.py
@@ -105,3 +105,17 @@ class GetSerialPortsError(PlatformioException):
 class GetLatestVersionError(PlatformioException):
 
     MESSAGE = "Can't retrieve latest PlatformIO version"
+
+
+class APIRequestError(PlatformioException):
+
+    MESSAGE = "[API] %s"
+
+
+class LibAlreadyInstalledError(PlatformioException):
+    pass
+
+
+class LibNotInstalledError(PlatformioException):
+
+    MESSAGE = "Library '%s' has not been installed yet"